+++ CHANGELOG.md
@@ -6,6 +6,11 @@ this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Refactored
+- Set the security headers as a global dictionary. Response security headers will be set via a loop.
+
 ## [0.8.0] - 2026-06-16
 
 ### Added